Fix buzz race condition
Previously, the question reader timer had a race condition with the buzz
action, where the timer appeared to tick after the user had buzzed. This
was because the Reader component was using useEffect to reactively
update the visible index whenever useReader updated the visible index
(which is an async operation). This was fixed by exposing a callback
that allowed the visible index to be directly updated by useReader
rather than reactively updating it in the Reader component.